feat: payingWallet
