+++ p2p_wallet/Scenes/Main/SendToken/Subscenes/ChooseRecipientAndNetwork/SendToken.ChooseRecipientAndNetwork.ViewModel.swift
@@ -40,6 +40,7 @@ extension SendToken.ChooseRecipientAndNetwork {
         private let navigationSubject = BehaviorRelay<NavigatableScene?>(value: nil)
         let recipientSubject = BehaviorRelay<SendToken.Recipient?>(value: nil)
         let networkSubject = BehaviorRelay<SendToken.Network>(value: .solana)
+        let payingWalletSubject = BehaviorRelay<Wallet?>(value: nil)
         
         // MARK: - Initializers
         init(
@@ -117,6 +118,7 @@ extension SendToken.ChooseRecipientAndNetwork.ViewModel: SendTokenChooseRecipien
     func save() {
         sendTokenViewModel.selectRecipient(recipientSubject.value)
         sendTokenViewModel.selectNetwork(networkSubject.value)
+        sendTokenViewModel.payingWalletSubject.accept(payingWalletSubject.value)
     }
     
     func navigateNext() {
